requirements.txt
Flask==3.0.2
Flask_Cors==4.0.0
Flask_RESTful==0.3.10
python-dotenv==1.0.1
torch==2.2.1
transformers==4.38.2
